Split network validation
