+++ src/game_data_validation_internal.h
@@ -107,6 +107,10 @@ bool require_unique_editor_stable_id(validation_context *ctx, name_table *stable
                                      const char *json_path);
 bool validation_mouse_button_name_valid(const char *name);
 bool validate_data_condition(validation_context *ctx, yyjson_val *condition, const char *path, validation_names *names);
+bool validate_action_array(validation_context *ctx, yyjson_val *actions, const char *json_path,
+                           validation_names *names);
+bool validate_network_port_value(validation_context *ctx, yyjson_val *value, const char *json_path, const char *label);
+bool validate_network(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool collect_input_actions(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool collect_input_assignment_sets(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool collect_input_profiles(validation_context *ctx, yyjson_val *root, validation_names *names);
